Implement JTAG simulation passthrough
This is a testing feature. The verilog simulator can now take a parameter
+jtag_port that will listen for a TCP connection. Messages sent to this
port will be driven over a JTAG interface to the device under test.

I created a TAP controller, but just wired it up in the top level testbench.
This can send and receive data, although there is a bug in sending data
from the target to the host where the LSB gets dropped.